Moved App Memory Tracker into it's own file
